+++ tests/fixtures/database/migrations/2022_10_11_144934_create_contacts_table.php
@@ -6,6 +6,11 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
     public function up()
     {
         Schema::create('Contacts', function (Blueprint $table) {
@@ -15,8 +20,13 @@ return new class extends Migration
         });
     }
 
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
     public function down()
     {
-        Schema::dropIfExists('contacts');
+        Schema::dropIfExists('Contacts');
     }
 };